ADS2 upd2

+++ src/alg.cpp
@@ -3,7 +3,7 @@
 
 
 double pown(double value, uint16_t n) {
-    double  result = 1; int i = 1;
+    double  result = 1; uint64_t i = 1;
     while (i <= n) {
         result *= value;
         i += 1;
@@ -12,7 +12,7 @@ double pown(double value, uint16_t n) {
 }
 
 uint64_t fact(uint16_t n) {
-    int result = 1, i = 1;
+    uint64_t result = 1, i = 1;
     while (i <= n) {
         result *= i;
         i += 1;
@@ -26,7 +26,7 @@ double calcItem(double x, uint16_t n) {
 
 double expn(double x, uint16_t count) {
     double result = 0;
-    for (int i = 0; i <= count; i++) {
+    for (uint64_t i = 0; i <= count; i++) {
         result += calcItem(x, i);
     }
     return result;
@@ -34,7 +34,7 @@ double expn(double x, uint16_t count) {
 
 double sinn(double x, uint16_t count) {
     double result = 0;
-    for (int i = 1; i <= count; i++) {
+    for (uint64_t i = 1; i <= count; i++) {
         result += calcItem(x, (2 * i) - 1) * pown(-1, i-1);
     }
     return result;
@@ -42,7 +42,7 @@ double sinn(double x, uint16_t count) {
 
 double cosn(double x, uint16_t count) {
     double result = 0;
-    for (int i = 1; i <= count; i++) {
+    for (uint64_t i = 1; i <= count; i++) {
         result += calcItem(x, (2 * i) - 2) * pown(-1, i-1);
     }
     return result;
